refactor: rename to feature_flags
